Moved ObservableQuery and WatchQueryOptions out of QueryManager
